model part finished.

+++ src/dao/interfaces/UserDao.java
@@ -2,10 +2,12 @@ package dao.interfaces;
 
 import dao.model.User;
 
+import java.util.List;
+
 public interface UserDao {
     boolean addUser(User u);
     boolean updateUser(User u);
     boolean deleteUser(User u);
     User getUserByUserName(String userName);
-    User getUserByUserId(int id);
+    User getUserByUserId(long id);
 }
